Allow flatpak to work with `-e qml`
Getting this to work required using the KDE flatpak platform to provide
qmlscene and QT modules as well as adding the QT WebEngine baseapp
because it is not packaged in the KDE platform.

I have tested that both the `qml` and `electron` versions of the app work,
but cannot test `ut` or `me`.

If these don't need to be available in the flatpak, this commit will
be fine by itself. Otherwise, I would recommend having an Axolotl-qml
flatpak which uses this build configuration.