Split flatpak manifests
Rather than have all launch options in one large flatpak file, this
provides two far smaller files.

The content of flatpak/web/ is the prior content of flatpak/. The
content of flatpak/qt/ is the prior conent of the -e qml version of the
flatpak.

Additionally, the qml option was replaced with qt, as this fits better
with what is actually happening. The original ut option also uses qml.