Fix the opening of HTTP links on Ubuntu Touch
When clicking on an HTTP link in a message it happened in some
situations that several tabs were opened in the web browser.

The reason is the event handler was registered each time a conversation
was opened, outside of the Vue event handling, and so wasn't
unregistered automatically.

The event is now handled in Vue and only one tab is opened.

See: /nanu-c/axolotl#718