Display a cross to close a fullscreen picture
When viewing a picture or video in fullscreen mode the "close" button
had no cross sign.
The "X" char was there but in white on a white background, so it was not
visible.
Its color has been set to black and instead of a "X" char the "times"
Font Awesome icon has been used.

See: /nanu-c/axolotl#720