Don't increment unread counter for current chat
When receiving a message while being on the conversation, the counter
of unread messages incremented and when coming back on the conversations
list there was a counter about an unread message.

It a message comes on the current conversation it should be considered
read so there is no need to increment the unread counter.

We now check the conversation to avoid incrementing the counter for the
current one.

Thanks to Blackoverflow for the help.

See /nanuc/axolotl#741