Adjust startup log message
As this message is shown on every startup, also on devices which runs neither Ubuntu nor Ubuntu Touch (like my system).

I would argue that the project has grown and matured, and to instead of explicitly mentioning Ubuntu the log message, slim down the message to focus on the application itself, rather than the system it runs on (or not).