Add the feature to send an attachment
The send message handler has been partially moved to messages.rs
to wrap the logic in a function send_message(). By doing so, both
the send message and the send attachment handlers use the same
function and avoid duplicate code.

In the future, it should be easy to change the send attachment feature
to:
1. Wait for the user to confirm before sending the attachment.
2. Add text with the attachment.

On the front end, the button to add an attachment has no style, it's just
a default button.