Transmit the message attachment to the front-end
The attachment's content type (audio, image, file) and the
filename (the attachment_id from Signal) are sent to the front-end.
With a HTTP server (todo) and some changes in the front-end (todo), the
attachment could be displayed.

Note: only the first attachment of the message is used currently. It's
possible for someone using Signal to send several pictures in one
message, so some adjustments will be done to show all the attachments.